refactor(api): move console running logic to dedicated class
