+++ sln/src/NSpec/Api/Controller.cs
@@ -1,12 +1,8 @@
 ﻿using Newtonsoft.Json;
 using NSpec.Api.Discovery;
 using NSpec.Api.Execution;
-using NSpec.Domain;
-using NSpec.Domain.Formatters;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 
 namespace NSpec.Api
 {
@@ -19,11 +15,10 @@ namespace NSpec.Api
             IDictionary<string, string> formatterOptions,
             bool failFast)
         {
-            var formatter = FindFormatter(formatterClassName, formatterOptions);
+            var batchExampleRunner = new BatchExampleRunner(testAssemblyPath,
+                tags, formatterClassName, formatterOptions, failFast);
 
-            var invocation = new RunnerInvocation(testAssemblyPath, tags, formatter, failFast);
-
-            int nrOfFailures = invocation.Run().Failures().Count();
+            int nrOfFailures = batchExampleRunner.Start();
 
             return nrOfFailures;
         }
@@ -62,42 +57,6 @@ namespace NSpec.Api
             var exampleRunner = new ExampleRunner(testAssemblyPath, onDiscovered, onExecuted);
 
             exampleRunner.Start(exampleFullNames);
-        }
-
-        /// <summary>
-        /// Find an implementation of IFormatter with the given class name
-        /// </summary>
-        /// <param name="formatterClassName"></param>
-        /// <param name="formatterOptions"></param>
-        /// <returns></returns>
-        static IFormatter FindFormatter(string formatterClassName, IDictionary<string, string> formatterOptions)
-        {
-            // Default formatter is the standard console formatter
-            if (string.IsNullOrEmpty(formatterClassName))
-            {
-                var consoleFormatter = new ConsoleFormatter();
-                consoleFormatter.Options = formatterOptions;
-                return consoleFormatter;
-            }
-
-            Assembly nspecAssembly = typeof(IFormatter).GetTypeInfo().Assembly;
-
-            // Look for a class that implements IFormatter with the provided name
-            var formatterType = nspecAssembly.GetTypes().FirstOrDefault(type =>
-                (type.Name.ToLowerInvariant() == formatterClassName)
-                && typeof(IFormatter).IsAssignableFrom(type));
-
-            if (formatterType != null)
-            {
-                var formatter = (IFormatter)Activator.CreateInstance(formatterType);
-                formatter.Options = formatterOptions;
-                return formatter;
-            }
-            else
-            {
-                throw new TypeLoadException("Could not find formatter type " + formatterClassName);
-
-            }
         }
     }
 }
